// frontend/api/model/accounttype.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Velmo
 * Backend API for Velmo
 * OpenAPI spec version: 1.0
 */

export type Accounttype = (typeof Accounttype)[keyof typeof Accounttype]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Accounttype = {
    Checking: 'Checking',
} as const