// frontend/api/model/createAccountRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Velmo
 * Backend API for Velmo
 * OpenAPI spec version: 1.0
 */
import type { Accounttype } from './accounttype'

export interface CreateAccountRequest {
    account_type: Accounttype
    balance: number
    iban: string
    name: string
}